combined helper functions under utils
